[chore] #115 타입 변경

+++ core/model/src/main/java/org/sopt/model/category/Category.kt
@@ -3,5 +3,5 @@ package org.sopt.model.category
 data class Category(
   val categoryId: Long,
   val categoryTitle: String,
-  val toastNum: Int,
+  val toastNum: Long,
 )
